rename: restriction form -> restriction modal
