+++ src/sections/upload-dataset-files/uploaded-files-list/restriction-modal/RestrictionModal.module.scss
